代码告警修改

+++ utils/vcard/src/vcard_decoder.cpp
@@ -82,7 +82,7 @@ std::string VCardDecoder::GetVersion()
 
 std::string VCardDecoder::GetVersionFromFileUtils(const std::string &line, int index)
 {
-    int versionIndex = index + std::string(VCARD_TYPE_VERSION).length() + 1;
+    size_t versionIndex = index + std::string(VCARD_TYPE_VERSION).length() + 1;
     if (versionIndex < static_cast<int>(line.length())) {
         return line.substr(versionIndex);
     }
